Changed read offsets to `size_t`, and explicitly reject negative offsets. I am not sure how else best to deal with these.

+++ fuse/internal.h
@@ -15,7 +15,7 @@ filename(const char *path);
 
 int
 sread(const char *s,
-	char *buf, size_t size, off_t offset);
+	char *buf, size_t size, size_t offset);
 
 int
 bibfs_reload(struct bibfs_state *b);
